Use an orange scheme for the autocompletion menu instead of red
The color scheme was effectively not setting a value for PmenuThumb
before: Vim only cares about the background color defined for that group
and we were only setting the foreground color. This commit fixes that by
setting the background color. As mentioned in a comment, though, I've
kept the foreground color defined because older versions of the
vim-airline theme for biogoo refer to "the foreground color of
PmenuThumb" and (since Vim itself doesn't care what the foreground color
of that group is) there's no sense in breaking the vim-airline theme.
(The slight change to the green that's used for this is intentional.)